The first version of Smart Door Lock
This version was made for testing purpose, where we did some test for very basic NFC card detection and reading it's data, ten comparision from the predefined data. if it matches then a http request is sent to update this data to the Google Sheets Server, recording the time and the name of the NFC Tag Holder.